# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Lab Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  detail?: string;
}

export type Severity = (typeof Severity)[keyof typeof Severity];

export const Severity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  info: "info",
} as const;

export interface ScanRequest {
  /** Target URL (must be http or https) */
  url: string;
}

export interface ScanFinding {
  id: string;
  title: string;
  severity: Severity;
  category: string;
  description: string;
  evidence?: string;
}

export type ScanResultHeaders = { [key: string]: string };

export type ScanResultSummary = {
  critical: number;
  high: number;
  medium: number;
  low: number;
  info: number;
};

export interface ScanResult {
  target: string;
  scannedAt: string;
  finalUrl?: string;
  statusCode?: number;
  durationMs?: number;
  headers?: ScanResultHeaders;
  findings: ScanFinding[];
  summary: ScanResultSummary;
}

export type CveItemReferencesItem = {
  url: string;
  source?: string | null;
};

export interface CveItem {
  id: string;
  published?: string;
  lastModified?: string;
  description: string;
  cvssScore?: number | null;
  cvssSeverity?: string | null;
  cvssVector?: string | null;
  cweIds: string[];
  affectedProducts: string[];
  references: CveItemReferencesItem[];
  patchAvailable: boolean;
}

export interface CveSearchResult {
  keyword: string;
  totalResults: number;
  items: CveItem[];
}

export interface HashCheckRequest {
  hash: string;
}

export interface HashCheckResult {
  hash: string;
  identifiedTypes: string[];
  primaryType?: string | null;
  length: number;
  /** Number of times this hash appears in HIBP breach data (SHA-1 only). Null if hash type is not SHA-1. */
  pwnedCount?: number | null;
  pwnedChecked: boolean;
  notes?: string | null;
}

export type DashboardStatsRecentActivityItem = {
  kind: string;
  label: string;
  at: string;
  severity?: string | null;
};

export interface DashboardStats {
  scansRun: number;
  cveLookups: number;
  hashChecks: number;
  criticalFindings: number;
  recentActivity: DashboardStatsRecentActivityItem[];
}

export type SearchCvesParams = {
  keyword: string;
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};