# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cybersecurity Lab Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hash/check:
    post:
      operationId: checkHash
      tags: [hash]
      summary: Identify a hash and check if it appears in known breach data (HIBP k-anonymity)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashCheckRequest"
      responses:
        "200":
          description: Hash analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashCheckResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required: [error]
...
    HashCheckRequest:
      type: object
      properties:
        hash:
          type: string
      required: [hash]
    HashCheckResult:
      type: object
      properties:
        hash:
          type: string
        identifiedTypes:
          type: array
          items:
            type: string
        primaryType:
          type: string
          nullable: true
        length:
          type: integer
        pwnedCount:
          type: integer
          nullable: true
          description: Number of times this hash appears in HIBP breach data (SHA-1 only). Null if hash type is not SHA-1.
        pwnedChecked:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [hash, identifiedTypes, length, pwnedChecked]
...