# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Lab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
  detail?: string;
}