# lib/api-zod/src/generated/types/hashCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Lab Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface HashCheckResult {
  hash: string;
  identifiedTypes: string[];
  primaryType?: string | null;
  length: number;
  /** Number of times this hash appears in HIBP breach data (SHA-1 only). Null if hash type is not SHA-1. */
  pwnedCount?: number | null;
  pwnedChecked: boolean;
  notes?: string | null;
}